Removes unnecessary theme
Prior to this layout of the site was split between the scktt layout and the scktt site.
As the design of this site is not going to be shared, there's no need to have it separated.

Also, I wanted to know for sure if the themes in hugo were required--turns out, they're not!